@@ -1,1 +1,1 @@
-GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function _FairDelegateWidgetState(){const inner=_FairDelegateWidgetState.__inner__;if(this==__global__){return new _FairDelegateWidgetState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_FairDelegateWidgetState.prototype.ctor.apply(this,args);return this;}}_FairDelegateWidgetState.__inner__=function inner(){};_FairDelegateWidgetState.prototype={};_FairDelegateWidgetState.prototype.ctor=function(){Object.prototype.ctor.call(this);};;return _FairDelegateWidgetState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
+GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function _FairDelegateWidgetState(){const inner=_FairDelegateWidgetState.__inner__;if(this==__global__){return new _FairDelegateWidgetState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_FairDelegateWidgetState.prototype.ctor.apply(this,args);return this;}}_FairDelegateWidgetState.__inner__=function inner(){};_FairDelegateWidgetState.prototype={getData:function getData(){const __thiz__=this;with(__thiz__){return['1','2','3','4','5','6','7','8','9','10'];}},};_FairDelegateWidgetState.prototype.ctor=function(){Object.prototype.ctor.call(this);};;return _FairDelegateWidgetState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
