@@ -4,9 +4,12 @@
  * found in the LICENSE file.
  */
 
+import 'dart:io';
+
 import 'package:fair/fair.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'internal/flexbuffer/fair_js_decoder_http_decoder.dart';
 import 'state.dart';
@@ -89,9 +92,7 @@ class FairApp extends InheritedWidget with AppState {
   }
 
   static FairApp? of(BuildContext? context, {bool rebuild = false}) {
-    return rebuild
-        ? context?.dependOnInheritedWidgetOfExactType<FairApp>()
-        : context?.findAncestorWidgetOfExactType<FairApp>();
+    return rebuild ? context?.dependOnInheritedWidgetOfExactType<FairApp>() : context?.findAncestorWidgetOfExactType<FairApp>();
   }
 
   String? pathOfBundle(String tag) {
@@ -106,13 +107,22 @@ class FairApp extends InheritedWidget with AppState {
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
-    properties
-        .add(DiagnosticsProperty<Map<String, String>>('bundle', bundleAlias));
+    properties.add(DiagnosticsProperty<Map<String, String>>('bundle', bundleAlias));
   }
 
   static void runApplication(Widget app, {Map<String, IFairPlugin>? plugins}) {
     // WidgetsFlutterBinding.ensureInitialized();
     FairPluginDispatcher.registerPlugins(plugins);
-    Runtime().loadCoreJs().then((value) => runApp(app));
+
+    if(Platform.isAndroid){
+      var runtime = Runtime();
+      var basicChannel = runtime.getBasicChannel();
+      basicChannel.invokeMethod('jsLoadListener').then((value){
+        runtime.loadCoreJs().then((value) => runApp(app));
+      });
+
+    }else{
+      Runtime().loadCoreJs().then((value) => runApp(app));
+    }
   }
 }
